+++ NeuralNetworkCode/src/Stimulus/WhiteNoiseLinear.cpp
@@ -28,7 +28,7 @@ void WhiteNoiseLinear::LoadParameters(std::vector<std::string> *input) {
 		SplitString(&(*it), &name, &values);
 
 		if ((name.find("seed") != std::string::npos)) {
-			seed = std::stod(values.at(0));
+			seed = static_cast<unsigned int>(std::stod(values.at(0)));
 		}
 		else if ((name.find("meanCurrent") != std::string::npos)) {
 			s.values.clear();
@@ -36,8 +36,8 @@ void WhiteNoiseLinear::LoadParameters(std::vector<std::string> *input) {
 				s.values.push_back(std::stod(values.at(2 * i)));
 				s.values.push_back(std::stod(values.at(2 * i + 1)));
 			}
-			s.start_time = std::round(std::stod(values.at(2 * P)) / info->dt);
-			s.end_time   = std::round(std::stod(values.at(2 * P + 1)) / info->dt);
+			s.start_time = static_cast<int>(std::round(std::stod(values.at(2 * P)) / info->dt));
+			s.end_time = static_cast<int>(std::round(std::stod(values.at(2 * P + 1)) / info->dt));
 
 			meanCurrent.push_back(s);
 		}
@@ -47,8 +47,8 @@ void WhiteNoiseLinear::LoadParameters(std::vector<std::string> *input) {
 				s.values.push_back(std::stod(values.at(2*i)));
 				s.values.push_back(std::stod(values.at(2*i+1)));
 				}
-			s.start_time = std::round(std::stod(values.at(2 * P)) / info->dt);
-			s.end_time   = std::round(std::stod(values.at(2 * P + 1)) / info->dt);
+			s.start_time = static_cast<int>(std::round(std::stod(values.at(2 * P)) / info->dt));
+			s.end_time = static_cast<int>(std::round(std::stod(values.at(2 * P + 1)) / info->dt));
 			sigmaCurrent.push_back(s);
 		}
 	}
